Moved views into fixtures.py
